streaming-pipeline/requirements_docker.txt
kafka-python==2.0.2
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0